+++ src/routes/(pages)/what-we-believe/+page.js
@@ -1,0 +1,6 @@
+export function load() {
+	return {
+		title: 'What we believe',
+		caption: 'Image by Gerd Altmann on Pixabay',
+	};
+}